Fix implicit type conversions
This commit fixes some warnings about implicit type conversions reported
by Visual Studio.

+++ number_theory/prime.h
@@ -36,10 +36,10 @@ class Sieve {
     is_prime_[1] = false;
     // Use uint64_t to avoid multiplication overflow. Due to the time and space
     // complexity of the algorithm, uint64_t is large enough.
-    for (uint64_t i = 2; i * i <= num_limit_; ++i) {
+    for (uint64_t i = 2; i * i <= static_cast<uint64_t>(num_limit_); ++i) {
       if (!is_prime_[i])
         continue;
-      for (uint64_t j = i * i; j <= num_limit_; j += i) {
+      for (uint64_t j = i * i; j <= static_cast<uint64_t>(num_limit_); j += i) {
         is_prime_[j] = false;
       }
     }
